Removed ListViewEx interface
Removed the interface that ListViewEx required and used inheritance instead: ListViewEx now has a generic parametre and is expected to be inherited to be used in Xaml.  This seems reasonable since our listviews in XAML are always bound to a specific type.

Also added a few context menu commands for TracklistControl.